third_party/WebKit/LayoutTests/fast/html/imports/import-dynamic-shared-expected.txt

CONSOLE WARNING: HTML Imports is deprecated and will be removed in M73, around March 2019. Please use ES modules instead. See https://www.chromestatus.com/features/5144752345317376 for more details.
This test ensures that de-duped import can be dynamically added.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS linkEl.import is harness.import
PASS successfullyParsed is true

TEST COMPLETE
